scalafmt: move ScalaFmtPreFormatProcessor from processors/scalafmt to scalafmt/processors package
